move Keys class to test helpers for ease of reuse
